fix : 분석 삭제하기 API 버그 수정
searchAnalysisWithVoiceOfAnswer 메서드에서 Answer이 없을 때도 가져와 질 수 있게 변경 -> 답변 없는 분석도 삭제 가능
AnswerNotFoundExcetion 커스텀 에외 생성 -> 사용할 필요 없어서 생성만 함

+++ src/main/java/com/project/doongdoong/domain/analysis/service/AnalysisServiceImp.java
@@ -4,6 +4,7 @@ import com.project.doongdoong.domain.analysis.dto.response.*;
 import com.project.doongdoong.domain.analysis.exception.AnalysisNotFoundException;
 import com.project.doongdoong.domain.analysis.model.Analysis;
 import com.project.doongdoong.domain.analysis.repository.AnalysisRepository;
+import com.project.doongdoong.domain.answer.exception.AnswerNotFoundException;
 import com.project.doongdoong.domain.answer.model.Answer;
 import com.project.doongdoong.domain.question.model.Question;
 import com.project.doongdoong.domain.question.model.QuestionContent;
@@ -234,8 +235,10 @@ public class AnalysisServiceImp implements AnalysisService{
         Analysis findAnalysis = analsisRepository.searchAnalysisWithVoiceOfAnswer(analysisId).orElseThrow(() -> new AnalysisNotFoundException());
         List<String> accessUrls = findAnalysis.getAnswers().stream().map(answer -> answer.getVoice().getAccessUrl())
                 .collect(Collectors.toList());
-        findAnalysis.getAnswers().stream().forEach(answer -> answer.disconnectWithVoice()); // 연관관계 끊기
-        voiceService.deleteVoices(accessUrls); // voice를 참조하는 객체 없으므로 삭제 가능 -> 벌크 삭제로 쿼리 최적화 필요
+        if(findAnalysis.getAnswers().size() != 0){
+            findAnalysis.getAnswers().stream().forEach(answer -> answer.disconnectWithVoice()); // 연관관계 끊기
+            voiceService.deleteVoices(accessUrls); // voice를 참조하는 객체 없으므로 삭제 가능 -> 벌크 삭제로 쿼리 최적화 필요
+        }
         // analysis 삭제로 question, answer 삭제 로직 -> voiceService.deleteVoices로 answer.voice와 관련 S3 파일은 이미 삭제.
         analsisRepository.deleteById(analysisId);
     }
